Custom keymaps all in one file
Even thought keymaps are used for various plugins configurations, I try
to keep them all in a single file: `lua/user/keymaps.lua`. This file can
be thought as a reference for the all custom defined keymaps. If new
keymaps are defined in the future I'll try to add them to `keymaps.lua`
file.