Telescope for "fuzzy finder" project navigation
Another way to navigate a project is by using a fuzzy finder: simply
type whatever you remember about file location/name and you get results
sorted by relevance. This way of navigation is the suggested one if you
have a mental model of the structure of the project and you just want to
reach a precise file quickly. `telescope-fzy-native` is added on top of
Telescope to dramatically improve sorting time and make the overall
experience more responsive.