Language Sever Protocol
Using a Language Sever Protocol (lsp) is the best way to enrich a simple
text editor (which support lsp) with functionalities that are shipped
with a full-fledged IDE. Go to definition, automatic formatting,
organize imports, show/go to references just to name a few.

Neovim (the client) talks to a program running in background (the sever)
using a protocol. The background program keep analyzing the code you're
typing and expose an API used by Neovim. The server it's language
specific (e.g. if you want to enhance Neovim with IDE features while
writing python code, `python-lsp-server` must be run in the background.
However you don't have to do it manually, this configuration automate all
aspect of lsp: installation, configuration and server management).
To enable lsp for a specific language simply open the source file and
run `:LspInstall`. Then restart Neovim et voilà.