LaTeX: lsp, compilation, forward/backward search
When you write tex/latex document it's convenient to always display the
pdf preview on one side of the screen while writing in the text editor
on the other. I like that changes that I made in .tex file quickly show
up in the pdf preview, so I enable the compilation on file save.

I use tectonic (https://github.com/tectonic-typesetting/tectonic) for
compilation and Skim (https://sourceforge.net/projects/skim-app/) for
pdf preview.

Another important feature is the synchronization between tex source and
the section of the pdf shown by Skim; this is achieved with forward and
backward search.
 - Use `:TexlabForward` command from Neovim to perform forward search
 - In Skim Cmd+Shift+click will perform backward search

There are a lot of moving parts in configuring Neovim to write LaTeX
docs so expect some frustration along the way, but the effort is worth
it.